Revert "feat(api): Allow configuring the API URL at runtime"
This reverts commit 0932ce5243556abe0f5434b19d5c32b71d7d7f2e.
