+++ src/api/index.ts
@@ -1,2 +1,2 @@
 export * from './apiService';
-export * from './requestService';
+export * from './request';
