+++ src/kogito-integration/KaotoEditorFactory.ts
@@ -8,13 +8,10 @@ import {
 import { KaotoEditorView } from "./KaotoEditorView";
 
 export class KaotoEditorFactory implements EditorFactory<Editor, KogitoEditorChannelApi> {
-
-  constructor(private readonly apiUrl?: string) {}
-
   public createEditor(
     envelopeContext: KogitoEditorEnvelopeContextType<KogitoEditorChannelApi>,
     initArgs: EditorInitArgs
   ): Promise<Editor> {
-    return Promise.resolve(new KaotoEditorView(envelopeContext, initArgs, this.apiUrl));
+    return Promise.resolve(new KaotoEditorView(envelopeContext, initArgs));
   }
 }
